CopyPaste: FreeImage_Copy's 1- and 4-bit row offsets were still 32-bit
AUDIT-TOOLKIT.md finding 28.  Upstream FreeImage 3.18.0 has the same defect one
power of two earlier; 0a8e25e halved it rather than introducing it, and the
first audit pass's "local" label was wrong.

    if (bpp == 1) {
        BOOL value;
        unsigned y_src, y_dst;              // 569

        for (int y = 0; y < dst_height; y++) {
            y_src = y * src_pitch;          // INT64 product, truncated into an unsigned

0a8e25e widened src_width, dst_width, dst_line, dst_pitch and src_pitch to INT64
and the 4-bit loop counters with them, but left the two row offsets alone.  The
product is computed in 64 bits and then truncated, so a 1- or 4-bit image whose
pixel data passes 4 GiB copies from and to the wrong rows.  3.18.0 declares
src_pitch as int, so there the whole product is 32-bit and it breaks at 2 GiB,
with signed overflow rather than truncation.

At a pitch of 2^28 the seventeenth row starts at exactly 2^32, so y_src comes
out 0.  A real 2147483647 x 17 1-bit bitmap - 4.25 GiB, peak RSS 4.46 GB, no
internal entry point involved:

    before  28p: copy scanline 16 = 00 00   (source has ff ff)   <-- took row 0 instead
    after   28p: copy scanline 16 = ff ff   (source has ff ff)

FreeImage_Copy returned a valid bitmap either way; only the pixels were wrong.
The 4-bit branch behaves identically (probes 28 and 28b use a sparse
MAP_NORESERVE mapping to get the same stride for a few pages of memory).  The
bpp >= 8 branch below was already INT64 throughout.

The two offsets are now INT64, and the 1-bit loop counters with them -
dst_height and dst_width have been INT64 since 0a8e25e, so the int counters were
narrowing those as well.

Verified: probes 28, 28b and 28p in .claude/audit/toolkit/tk2.c.  All 2281
regression cases are byte-identical to the parent commit.

+++ Source/FreeImageToolkit/CopyPaste.cpp
@@ -566,12 +566,18 @@ FreeImage_Copy(FIBITMAP *src, int left, int top, int right, int bottom) {
 	// copy the bits
 	if (bpp == 1) {
 		BOOL value;
-		unsigned y_src, y_dst;
+		// y * src_pitch is an INT64 product - 0a8e25e widened the pitches - but
+		// it was stored in an unsigned and truncated there, so a 1- or 4-bit
+		// image whose pixel data passes 4 GiB copied from and to the wrong rows.
+		// 3.18.0 declares src_pitch as int and breaks at 2 GiB instead, which is
+		// what this replaces rather than introduces.  The loop counters go with
+		// them: dst_height and dst_width are INT64 above.
+		INT64 y_src, y_dst;
 
-		for (int y = 0; y < dst_height; y++) {
+		for (INT64 y = 0; y < dst_height; y++) {
 			y_src = y * src_pitch;
 			y_dst = y * dst_pitch;
-			for (int x = 0; x < dst_width; x++) {
+			for (INT64 x = 0; x < dst_width; x++) {
 				// get bit at (y, x) in src image
 				value = (src_bits[y_src + ((left + x) >> 3)] & (0x80 >> ((left + x) & 0x07))) != 0;
 				// set bit at (y, x) in dst image
@@ -582,7 +588,7 @@ FreeImage_Copy(FIBITMAP *src, int left, int top, int right, int bottom) {
 
 	else if (bpp == 4) {
 		BYTE shift, value;
-		unsigned y_src, y_dst;
+		INT64 y_src, y_dst;
 
 		for (INT64 y = 0; y < dst_height; y++) {
 			y_src = y * src_pitch;
